py-backend/requirements.txt
fastapi
uvicorn
pymongo
face_recognition
dlib
opencv-python
numpy
python-dotenv
